Refactor Arquillian extensions
This change separated the Arquillian extension behaviours
into their own separate classes for readability.
