+++ appserver/tests/tck/microprofile/config/src/main/java/org/glassfish/microprofile/config/tck/ConfigArquillianExtension.java
@@ -15,46 +15,20 @@
  */
 package org.glassfish.microprofile.config.tck;
 
+import org.glassfish.microprofile.config.tck.client.BeansXmlTransformer;
+import org.glassfish.microprofile.config.tck.client.ConfigDeploymentExceptionTransformer;
+import org.glassfish.microprofile.config.tck.client.LibraryIncluder;
+import org.glassfish.microprofile.config.tck.remote.ConfigRemoteArquillianExtension;
 import org.jboss.arquillian.container.spi.client.container.DeploymentExceptionTransformer;
 import org.jboss.arquillian.container.test.spi.client.deployment.ApplicationArchiveProcessor;
+import org.jboss.arquillian.container.test.spi.client.deployment.AuxiliaryArchiveAppender;
 import org.jboss.arquillian.core.spi.LoadableExtension;
-import org.jboss.arquillian.test.spi.TestClass;
-import org.jboss.shrinkwrap.api.Archive;
-import org.jboss.shrinkwrap.api.ArchivePath;
-import org.jboss.shrinkwrap.api.asset.ArchiveAsset;
-import org.jboss.shrinkwrap.api.asset.UrlAsset;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.jboss.shrinkwrap.resolver.api.maven.Maven;
 
-import java.io.File;
-import java.net.URL;
-import java.util.function.Consumer;
 import java.util.logging.Logger;
 
-import static java.lang.String.format;
-import static java.util.logging.Level.SEVERE;
-
-/**
- * This extension performs the following duties for TCK runs:
- *  - Adding Hamcrest to each deployment, to prevent ClassNotFoundExceptions when running hamcrest tests
- *  - Replacing beans.xml files with ones declaring the 'all' bean discovery type.
- *    This is because version 3.0.1 of the TCK still deploys an empty beans.xml due to a faulty assumption that
- *    CDI < 4 is still defaulting to the 'all' type.
- */
-public class ConfigArquillianExtension implements LoadableExtension, ApplicationArchiveProcessor {
+public class ConfigArquillianExtension implements LoadableExtension {
 
     private static final Logger LOGGER = Logger.getLogger(ConfigArquillianExtension.class.getName());
-    private static final String BEANS_XML_PATH = format("/META-INF%sbeans.xml", ArchivePath.SEPARATOR);
-    private static final String LIB_DIR_PATH = format("WEB-INF%slib", ArchivePath.SEPARATOR);
-
-    private final URL beansXmlResource;
-
-    public ConfigArquillianExtension() {
-        this.beansXmlResource = getClass().getClassLoader().getResource("beans.xml");
-        if (beansXmlResource == null) {
-            throw new IllegalStateException("Unable to find beans.xml resource in test dir");
-        }
-    }
 
     /**
      * Register this object as an Arquillian extension
@@ -62,54 +36,11 @@ public class ConfigArquillianExtension implements LoadableExtension, Application
      */
     @Override
     public void register(ExtensionBuilder extensionBuilder) {
-        extensionBuilder.service(ApplicationArchiveProcessor.class, getClass());
+
+        LOGGER.info("Client Arquillian extension registered");
+
+        extensionBuilder.service(ApplicationArchiveProcessor.class, BeansXmlTransformer.class);
+        extensionBuilder.service(ApplicationArchiveProcessor.class, LibraryIncluder.class);
         extensionBuilder.service(DeploymentExceptionTransformer.class, ConfigDeploymentExceptionTransformer.class);
-    }
-
-    @Override
-    public void process(Archive<?> archive, TestClass testClass) {
-        if (!(archive instanceof WebArchive)) {
-            return;
-        }
-        replaceBeansXml(archive);
-        addDependencies((WebArchive) archive);
-    }
-
-    private void replaceBeansXml(Archive<?> archive) {
-        final var beansXml = archive.get(BEANS_XML_PATH);
-        if (beansXml != null) {
-            LOGGER.info(() -> format("Replacing beans.xml in archive [%s]", archive.getName()));
-            archive.add(new UrlAsset(beansXmlResource), BEANS_XML_PATH);
-        }
-        processLibraries(archive, this::replaceBeansXml);
-    }
-
-    private void addDependencies(WebArchive archive) {
-        try {
-            archive.addAsLibraries(resolveDependency("org.hamcrest:hamcrest:2.2"));
-        } catch (Exception e) {
-            LOGGER.log(SEVERE, "Error adding dependencies", e);
-        }
-    }
-
-    private static void processLibraries(Archive<?> archive, Consumer<Archive<?>> consumer) {
-        final var libDir = archive.get(LIB_DIR_PATH);
-
-        if (libDir != null) {
-            libDir.getChildren()
-                    .forEach(node -> {
-                        final var asset = node.getAsset();
-                        if (asset instanceof ArchiveAsset) {
-                            LOGGER.info(() -> format("Processing subarchive [%s]", node.getPath()));
-                            consumer.accept(((ArchiveAsset) asset).getArchive());
-                        }
-                    });
-        }
-    }
-
-    private static File[] resolveDependency(String coordinates) {
-        return Maven.resolver()
-                .resolve(coordinates)
-                .withoutTransitivity().asFile();
     }
 }
